Refactor and enhance various components; as per ruff
- Updated security.py to patch bcrypt compatibility for passlib.
- Cleaned up whitespace in main.py.
- Added LLMUsageLog to the models' __init__.py exports.
- Improved type checking in credential_vault.py, refresh_token_session.py, resume_profile.py, and search_preference.py.
- Consolidated created_at column definition in llm_usage_log.py.
- Refactored PipelineCheckpointer to improve type casting and error handling.
- Enhanced logging in base.py for missing fields in scraped data.
- Updated email_monitor_service.py to streamline message classification methods.
- Improved usage_service.py for logging LLM token usage.
- Cleaned up seed_e2e_data.py for better readability and consistency.
- Adjusted tests to improve monkeypatching and assertions.

+++ backend/app/services/profile_completeness_service.py
@@ -24,7 +24,7 @@ class ProfileCompletenessService:
             }
 
             missing_fields = [field for field, present in checks.items() if not present]
-            score = int(round((len(checks) - len(missing_fields)) / len(checks) * 100))
+            score = round((len(checks) - len(missing_fields)) / len(checks) * 100)
 
             recommendations = [
                 f"Add or improve {field.replace('_', ' ')} to strengthen job matching." for field in missing_fields
